third_party/WebKit/LayoutTests/inspector-protocol/css/css-set-property-text-inline-expected.txt

Some red text goes here
==== Initial style sheet text ====
color: red; border: 0px;

Running test: testEditProperty
==== Style sheet text ====
content: 'EDITED PROPERTY'; border: 0px;

Running test: testBreakingCommentEditProperty
Expected protocol error: SyntaxError The property '/*<--OPENED COMMENT' could not be set.

Running test: testInsertFirstProperty
==== Style sheet text ====
content: 'INSERTED PROPERTY';color: red; border: 0px;

Running test: testInsertMultipleProperties
==== Style sheet text ====
color: red; content: 'INSERTED #1';content: 'INSERTED #2';border: 0px;

Running test: testInsertLastProperty
==== Style sheet text ====
color: red; border: 0px;content: 'INSERTED PROPERTY';
